github: Expand OAuth scope to read:org
Supersedes #983
Resolves #972

The new --reviewer flag allows specifying teams as reviewers
in the form 'org/team'.
To resolve team reviewers, we need to use:

    organization(login: $org) {
      team(slug: $teamSlug) {
        id
      }
    }

This requires the `read:org` OAuth scope.

Additionally, when reading information about a PR
that has a team reviewer assigned to it, we use:

    $pullRequestNode {
      reviewRequests(..) {
        nodes {
          requestedReviewer {
             ... on Actor {
               login
          }
        }
      }
    }

Which, again, requires the `read:org` scope to resolve team reviewers.

While it's possible to hack around this with the v3 API (#983),
that's just adding more complexity to the codebase.

After discussing with a couple users,
it sounds like `read:org` is an acceptable scope to request.
So we'll just expand scope and stop trying to hack around it.
