+++ .changes/v0.21.0.md
@@ -9,7 +9,7 @@ This unlocks a couple common use cases:
 
 ```bash
 # Always request reviews from my team
-git config spice.submit.reviewers '@example/my-team'
+git config spice.submit.reviewers 'example/my-team'
 
 # Always assign submitted CRs to myself
 git config spice.submit.assignees $(whoami)
